* amd64fbsd-nat.c (_initialize_amd64fbsd_nat): Change type of ps_strings into a long.

+++ gdb/amd64fbsd-nat.c
@@ -289,7 +289,7 @@ Please report this to <bug-gdb@gnu.org>.",
      environment can be found.  */
   {
     int mib[2];
-    int ps_strings;
+    long ps_strings;
     size_t len;
 
     mib[0] = CTL_KERN;
